oficios-automation/funcoes/W1_ingestao_trigger/requirements.txt
google-cloud-firestore>=2.14.0
google-cloud-pubsub>=2.18.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
groq>=0.4.0
